Move workflow definition tests to workflow module
